Replace NoodleKit's marker implementation with a new one. The delegate protocol is the same, for now.
Since the original author mistook NSScrollView's document view for an NSDocument, the fileName parameter was useless, so we always pass nil in the new implementation to retain compatibility. Also, we don't check if the delegate responds to the protocol's current methods, because they are both mandatory, and if their implementation is missing we should crash, not silently fail.