Improve ruler tearing again by using faster APIs.
rectArrayForCharacterRange:withinSelectedCharacterRange:inTextContainer:rectCount: is slow because it wastes a lot of time checking if the character range is in the selection or near the insertion point, but lineFragmentRectForGlyphIndex:effectiveRange: does not and is much much faster. boundingRectForGlyphRange:inTextContainer: checks for vicinity of the insertion point too, but it returns a single rect instead of an array, and should be a tad faster; doesn't matter though, since we call it only if we can't use lineFragmentRectForGlyphIndex:effectiveRange; I switched to it for this marginal case because it removes the need for making a rect array.

Still tears a bit though...