Remove non-working menu constructors and associated action methods.
This code relied on proper behavior of SMLBasicPerformer, a leftover from Smultron which is now a dummy class (this code was the only thing that used it, and the used methods are basically empty). Since SMLBasicPerformer is not exported, an app using Fragaria could not even think of subclassing it, making it 100% useless.