Use CoreText for drawing the line numbers in the gutter. This makes the gutter not lag anymore 90% of the time.
I wanted to make this change for a while because while profiling, half the drawing time of the gutter view was spent drawing the text (this was before the addition of error display in the gutter). But I didn't know if Core Text's metrics were the same as Cocoa's metrics. Indeed they are a bit different (CT returns more precise figures), but I think I have nailed the right match of roundings to have the baselines match.

Just for the record, I looked at Xcode's gutter view in a debugger, and it turns out it implements the line number gutter in the same way, using an NSRulerView subclass and drawing the line numbers in it with core text.