Backwards-incompatible API change! Removed embedInView: in favor of initWithView:.
This is *radical*! I really really wanted to avoid this, but if we have to properly remove the docSpec, this has to be done.

Suppose I want to remove a property from the docSpec, and I want to move it to a class of Fragaria (like MGSSyntaxColouring). Since these accessory classes are created in embedInView: and the readme says to set the property I want to move before calling embedInView:, I can't move anything without silently breaking everything. Removing embedInView: still breaks everything but sooner (in the build phase), and the authors of Fragaria-using apps will not think this is a bug.