Reworked the MGSPreferences a little bit with warnings and deprecations. Current functionality is still identical, but we're nearing a point where we'll have to change implementation API again.
Right now Fragaria manages all of its own userDefaults. While this can be convenient for single instances in an app, some of the properties affect every instance in an app.

Additionally the containing app is responsible for managing all user defaults. However Fragaria does have a lot of defaults, and so MGSFragariaPreferences will eventually be transitioned to its own Framework that provides defaults services to Fragaria if wanted. Otherwise, Fragaria is completely property-driven and no longer tied to NSUserDefaults.