Update syntax colouring from drawRect: instead of listening for notifications.
This improves scrolling performance *a lot* because the syntax colouring now keeps in sync with NSTextView's tile cache. Since the colouring is still done on the main thread, nothing else is slowed down. Note that removing all colours from the layout manager causes a redraw, and this makes unnecessary all manual refresh of the colouring data when the colouring options change.