Calling -selectedFont: inside of -changeFont: may return nil (and is strongly discouraged by the documentation).
The convertFont method of NSFontManager is Cocoa's worst named method. NSFontManager tracks the changes to selected font made in the font panel and, when convertFont is called, it applies the changes to the specified font. So, passing nil to convertFont will return nil, screwing everything up.

This nil-font bug situation could be reproduced by opening the font panel and then the fonts and colours view controller, and after that change the font without clicking the "Change Font" button.