Rename showsWarningsInGutter to showsSyntaxErrors in MGSFragaria.
Also remove useless todo. If we want KVO to work for this property, it must stay where it is, how it is. Moving it to MGSLineNumberView is not OK because the line number view only *shows* syntax errors, it should not *control* them.
What's wrong with this code anyway? It works :-)