When replacing text storage, set typing attributes from the text storage instead of setting them to the text storage.
Since the typing attributes in other views using that text storage are not updated when changing the text storage, the old behavior could screw up editing on the
other views. This way, instead, the "new" view effectively inherits the typing attributes from the other. This new behavior is equivalent to the old one if the textStorage passed in is empty.